% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapter2.R
\name{chapter2}
\alias{chapter2}
\title{Demonstrates what I did with the chapter 2 excercises.}
\usage{
chapter2(diamonds)
}
\arguments{
\item{diamonds, }{the diamonds input file}
}
\value{
a file containing my results.
}
\description{
Demonstrates what I did with the chapter 2 excercises.
}
\author{
Femke Thon
}

