% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapter3.R
\name{chapter3}
\alias{chapter3}
\title{Demonstrates what I did with the chapter 2 excercises.}
\usage{
chapter3(input)
}
\arguments{
\item{diamonds, }{the diamonds input file}
}
\value{
a file containing my results.
}
\description{
Demonstrates what I did with the chapter 2 excercises.
}
\author{
Femke Thon
}

